+++ fluent_utils/django_compat/__init__.py
@@ -1,7 +1,8 @@
 from .django14 import now, utc, patterns, url, include
 from .django15 import get_user_model, AUTH_USER_MODEL
 from .django16 import transaction_atomic, add_preserved_filters
-from .django17 import is_installed, get_models, get_app_names
+from .django17 import is_installed, get_models, get_app_names, get_app_label
+from .django18 import CallableChoiceIterator
 from .python3 import smart_text
 
 __all__ = (
@@ -10,6 +11,6 @@ __all__ = (
     'get_user_model', 'AUTH_USER_MODEL',
     'transaction_atomic',
     'add_preserved_filters',
-    'is_installed', 'get_models', 'get_app_names'
+    'is_installed', 'get_models', 'get_app_names', 'get_app_label',
     'smart_text',
 )