+++ fluent_utils/django_compat/django17.py
@@ -1,4 +1,6 @@
+import sys
 from django.conf import settings
+
 
 try:
     from django.apps import apps
@@ -11,6 +13,13 @@ except ImportError:
 
     def get_app_names():
         return settings.INSTALLED_APPS
+
+    def get_app_label(module):
+        # Django <= 1.6 does not know apps.
+        # Figure out the app_label by looking one level up.
+        # For 'django.contrib.sites.models', this would be 'sites'.
+        model_module = sys.modules[module]
+        return model_module.__name__.split('.')[-2]
 else:
     # Django 1.7 provides an official API, and INSTALLED_APPS may contain non-string values too.
     is_installed = apps.is_installed
@@ -20,3 +29,8 @@ else:
         return [
             appconfig.name for appconfig in apps.get_app_configs()
         ]
+
+    def get_app_label(module):
+        # Django 1.7 knows the "apps", so it can provide the meta information
+        app_config = apps.get_containing_app_config(module)
+        return app_config.label